refactor: Add docs, rename package
