+++ pkg/kubernetesclient/client.go
@@ -1,4 +1,4 @@
-package client
+package kubernetesclient
 
 import (
 	"context"
